+++ R/boot.cgaim.R
@@ -84,7 +84,7 @@
 #' boot2 <- boot.cgaim(ans, bsamples = bsamp)
 #' 
 #' # Parallel computing (two cores)
-#' \dontrun{
+#' \donttest{
 #' boot3 <- boot.cgaim(ans, nc = 2)
 #' }
 #'