+++ R/check_cmat.R
@@ -11,7 +11,7 @@ check_cmat <- function(Cmat){
   redundant <- rep(FALSE, ncons)
   for (i in seq_len(ncons)){
     y <- tCmat[,i]
-    x <- tCmat[,-c(i, which(redundant)),drop = F]
+    x <- tCmat[,-c(i, which(redundant)),drop = FALSE]
     res <- limSolve::nnls(x, y)
     redundant[i] <- isTRUE(all.equal(y, drop(x %*% res$X)))
   }